requirements.txt
playwright==1.40.0
openai>=1.10.0
python-dotenv==1.0.0
pydantic==2.5.0
rich==13.7.0
typer==0.9.0
beautifulsoup4==4.12.2
lxml==4.9.3
pillow==10.1.0
requests==2.31.0
httpx>=0.25.0
